asm-verify: fix signature mismatches incl. Touch float coords (#266)
From the signature-mismatch round (verified vs v1.6.1): the port carried several
function signatures that diverged from the binary, so asm-verify never paired
them. Fix the real ones:

- Mortar::Touch: coords were truncated to int at the storage site
  (___UpdateInternal cast SDL float coords to int32_t), losing sub-pixel blade
  precision. TouchState x/y -> float; GetTouchPos/GetTouchDelta -> (uint, float&,
  float&); GetTouchInReigion -> (float,float,float,float); InputDeviceBada locals
  -> float. (Touch::GetTouchPos @0x002429d4 / UpdateInternal @0x00242868.)
- MortarSound::SetPitch(unsigned int) -> (float) (@0x00230218; VFP ABI).
- Math::Random::Seed(uint64) -> (uint32) + reset m_Mult/m_Inc (@0x0012c998;
  verified no determinism change -- the only seed source is 32-bit).
- MatrixManager::SetupOrtho/SetupLookAt: re-add the defaulted out-matrix param
  the binary takes (@0x002571d0/@0x002572ac), matching SetupPerspective.
- SplatEntity::Init(): add the (void*,long,Vec3*) vtable args the binary declares
  (ignored, @0x001eb264).

Tests: test_utf8_text stub updated for the new MatrixManager signatures;
test_random seeds with a 32-bit value (ctor stays 64-bit / Seed is 32-bit, as in
the binary).

+++ src/engine/audio/MortarSound.cpp
@@ -156,10 +156,9 @@ bool MortarSound::IsReady() {
     return true;
 }
 
-// DIFFERS: v1.6.1 binary @ 0x0018c778 discards pitch arg; not implementing real pitch shift to match
-void MortarSound::SetPitch(unsigned int /*pitch*/) {
+// ASM-spec v1.6.1 MortarSound::SetPitch @0x00230218: (float) — body discards the pitch arg.
+void MortarSound::SetPitch(float /*pitch*/) {
     if (m_Handle == 0) m_State = 0;
-    // binary discards the pitch argument; not implementing real pitch shift to match
 }
 
 } // namespace Mortar
